.github: add CodeQL workflow
This is the automatically generated config file for CodeQL, a code
query language, which can be used to run bespoke static analysis
tasks.

It is currently set to "autobuild" mode, which in Github's terminology
means it tries `./configure && make`, which fails because configure
looks for packages that aren't installed.

In another commit or two we'll add the necessary bootstrap lines, and
perhaps remove some of the unneeded bits.

Signed-off-by: Douglas Bagnall <douglas.bagnall@catalyst.net.nz>
Reviewed-by: Martin Schwenke <martin@meltin.net>