vfs: Add SMB_VFS_RENAME_STREAM
[MS-FSA] 2.1.5.15.12.1 Algorithm for Performing Stream Rename

is simpler and diffent enough from renaming a file or directory that I
believe a separate VFS operation is justified instead of tunneling it
through the renameat call. For example it's only possible to rename
streams within a file, so only one stream open fsp and a newname which
is guaranteed to be a stream name is necessary.

Add stub implementations to our streams modules, to be filled later.

Signed-off-by: Volker Lendecke <vl@samba.org>

vfs_streams_xattr: rename_streams
Reviewed-by: Ralph Boehme <slow@samba.org>