ctdb-event: Don't replace an existing result with NULL
If script_list is NULL then event->script_list is set to NULL by
TALLOC_FREE().  This seems like the wrong thing to do because NULL
indicates a problem running the current event.  We should keep the
previous result because it contains useful information.

In theory, this can't happen because the caller checks for NULL.
However, given that the check is here, it might as well do the right
thing... and this simplifies a subsequent change.

Signed-off-by: Martin Schwenke <mschwenke@ddn.com>
Reviewed-by: Ralph Boehme <slow@samba.org>