ctdb-common: Make run_event return full script list
run_event currently truncates the resulting script list after the
first failure.

Instead, return the full script list, with scripts that weren't run
flagged with ENODATA in their summary.  This structure is zeroed on
allocation, so this is the only field that needs to be set.

For now, have the event tool skip such scripts.  Do this in
print_status() instead of print_status_one(), since there will soon be
additional checks in print_status() and it makes sense to put them
together.

Signed-off-by: Martin Schwenke <mschwenke@ddn.com>
Reviewed-by: Ralph Boehme <slow@samba.org>