lib: Add adouble_buf_parse()
Simplified version of ad_get that takes a buffer and does basic parsing of an
AppleDouble file format. The entries are represented as DATA_BLOBs directly
pointing at "buf" to avoid offset calculations in users of this.

Yes, this is a duplication of logic, but it makes the next patch
possible. Future patches could use this in ad_unpack()

Signed-off-by: Volker Lendecke <vl@samba.org>
Reviewed-by: Ralph Boehme <slow@samba.org>