mdssvc: reduce a log level to DEBUG
The expression

  InRange(*,$time.iso(2024-12-31T23:00:00Z),$time.iso(2025-12-31T23:00:00Z))

in a Spotlight query produces the following log message:

  map_fts: Mapping fts [757378800] unexpected op [~]

However, when

    elasticsearch:ignore unknown attribute = yes

is set, the parser will ignore the failed expression and continue
parsing given the expression is part of a larger expression like
"subexpression1 OR subexpression2". Avoid spamming the log and reduce
the loglevel when we hit this case.

BUG: https://bugzilla.samba.org/show_bug.cgi?id=15930

Signed-off-by: Ralph Boehme <slow@samba.org>
Reviewed-by: Volker Lendecke <vl@samba.org>
(cherry picked from commit 3b2b50983901ad5039124b82f149b30675c80a9a)