s3:printing: Load the shares for [printers] in samba-bgqd
One of the main functions of bgqd is:

        delete_and_reload_printers_full()

It isn't able to do its work, if we don't load the shares. Normally bgqd was
forked from smbd and this loaded the shares. But with the introduction of
samba-dcerpcd it is a standalone service now. As a standalone service it is
responsible to load the shares if it needs to work on them.

The following message is printed if delete_and_reload_printers_full() tries to
do its job:

[2025/10/23 09:57:27,  7, pid=41935, effective(0, 0), real(0, 0)] ../../source3/param/loadparm.c:4419(lp_servicenumber)
  lp_servicenumber: couldn't find printers
[2025/10/23 09:57:27,  7, pid=41935, effective(0, 0), real(0, 0)] ../../source3/param/loadparm.c:4419(lp_servicenumber)
  lp_servicenumber: couldn't find printers

BUG: https://bugzilla.samba.org/show_bug.cgi?id=15936

Signed-off-by: Andreas Schneider <asn@samba.org>
Reviewed-by: Anoop C S <anoopcs@samba.org>